AS-3984

+++ src/main/java/com/asg/shipchandling/deliverynote/dto/UpdateSalesDeliveryNoteRequest.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -22,7 +23,7 @@ public class UpdateSalesDeliveryNoteRequest {
     @Size(max = 20, message = "Currency code must not exceed 20 characters")
     private String currencyCode;
     
-    private Long currencyRate;
+    private BigDecimal currencyRate;
 
     @Size(max = 20, message = "Delivery status must not exceed 20 characters")
     private String deliveryStatus;
